+++ GameServer/ClientProxy.cpp
@@ -1,7 +1,7 @@
 #include "ClientProxy.h"
 
 
-ClientProxy::ClientProxy(int _id, std::string _alias, sf::IpAddress _ip, unsigned short _port, sf::Vector2i headPos)
+ClientProxy::ClientProxy(int _id, std::string _alias, sf::IpAddress _ip, unsigned short _port, sf::Vector2f headPos)
 {
 	id = _id;
 	alias = _alias;
@@ -9,32 +9,32 @@ ClientProxy::ClientProxy(int _id, std::string _alias, sf::IpAddress _ip, unsigne
 	port = _port;
 	numPings = 0;
 	CreateInitialBodyPositions(headPos);
-	accumMovement = sf::Vector2i(0,0);
+	accumMovement = sf::Vector2f(0,0);
 }
 
 ClientProxy::~ClientProxy()
 {
 }
 
-void ClientProxy::CreateInitialBodyPositions(sf::Vector2i headPos)
+void ClientProxy::CreateInitialBodyPositions(sf::Vector2f headPos)
 {
 	bodyPositions.push_back(headPos);
 
 	//create random direction
-	direction = sf::Vector2i(1.f,0.f);
+	direction = sf::Vector2f(1.f,0.f);
 
 	//make two more bodies:
-	sf::Vector2i pos = headPos + direction * separation;
+	sf::Vector2f pos = headPos + direction * separation;
 	bodyPositions.push_back(pos);
 
-	sf::Vector2i secondPos = pos + direction * separation;
+	sf::Vector2f secondPos = pos + direction * separation;
 	bodyPositions.push_back(secondPos);
 }
 
 void ClientProxy::PutBodyPositions(sf::Packet* pack)
 {
 	*pack << (int)bodyPositions.size();
-	for each (sf::Vector2i pos in bodyPositions)
+	for each (sf::Vector2f pos in bodyPositions)
 	{
 		*pack << pos.x;
 		*pack << pos.y;
@@ -48,24 +48,24 @@ void ClientProxy::AddDataToPacket(sf::Packet* pack)
 	PutBodyPositions(pack);
 }
 
-sf::Vector2i ClientProxy::SumToHeadPosition()
+sf::Vector2f ClientProxy::SumToHeadPosition()
 {
 	return bodyPositions[0] + accumMovement;
 }
 
 void ClientProxy::CreateBodyPosition()
 {
-	sf::Vector2i pos = bodyPositions.back();
+	sf::Vector2f pos = bodyPositions.back();
 	bodyPositions.push_back(pos);
 }
 
-void ClientProxy::UpdatePosition(sf::Vector2i headPos)
+void ClientProxy::UpdatePosition(sf::Vector2f headPos)
 {
 	bodyPositions[0] = headPos;
 
 	for (int i = 1; i < (int)bodyPositions.size() ; i++)
 	{
-		sf::Vector2i dir = bodyPositions[i] - bodyPositions[i - 1];
+		sf::Vector2f dir = bodyPositions[i] - bodyPositions[i - 1];
 		Normalize(dir);
 		bodyPositions[i] = bodyPositions[i - 1] + dir * separation;
 	}
@@ -76,5 +76,5 @@ void ClientProxy::UpdatePosition(sf::Vector2i headPos)
 	}
 	bodyPositions[0] = headPos;*/
 
-	accumMovement = sf::Vector2i(0,0);
+	accumMovement = sf::Vector2f(0,0);
 }
